# openapi.yaml
...
swagger: "2.0"
info:
  description: "Documentacion de la API de prueba de revibe"
  title: "revibe_23_06_21"
  version: "1.0.0"
# [END swagger]
consumes:
- "application/json"
produces:
- "application/json"
schemes:
# Uncomment the next line if you configure SSL for this API.
#- "https"
- "http"
paths:
...
  "/spotify/userLikes":
        get:
          description: "Funcion para obtener los favoritos de spotify."
          operationId: "userLikes"
          x-google-backend:
            address: https://us-central1-moodsic-proyecto.cloudfunctions.net/spotify-userLikes
          produces:
          - "application/json"
          responses:
            200:
              description: "userLikes"
              schema:
                type: string
            400:
              description: "not userLikes"
              schema:
                type: string
  "/spotify/userFeatures":
        get:
          description: "Funcion para obtener las userFeatures."
          operationId: "userFeatures"
          x-google-backend:
            address: https://us-central1-moodsic-proyecto.cloudfunctions.net/spotify-userFeatures
          produces:
          - "application/json"
          responses:
            200:
              description: "userFeatures"
              schema:
                type: string
            400:
              description: "not userFeatures"
              schema:
                type: string
...
  "/auth/callback":
          get:
            description: "Funcion callback para la autenticacion de Spotify."
            operationId: "callback"
            x-google-backend:
              address: https://us-central1-moodsic-proyecto.cloudfunctions.net/spotify-callback
            produces:
            - "application/json"
            responses:
              200:
                description: "Callback success"
                schema:
                  type: string
              400:
                description: "Error en al autententicación"
                schema:
                  type: string
  "/auth/refreshToken":
        put:
          description: "Refrescar el access token de Spotify"
          operationId: "refreshToken"
          x-google-backend:
            address: https://us-central1-moodsic-proyecto.cloudfunctions.net/spotify-refreshToken
          produces:
          - "application/json"
          responses:
            200:
              description: "Data encontrada"
              schema:
                type: string
         
            400:
              description: "Data no encontrada"
              schema:
                type: string
...